refactor: move document file type schemas
