+++ src/shared/schemas/index.ts
@@ -3,6 +3,7 @@ export * from './identifiers';
 export * from './agent';
 export * from './agentCliSettings';
 export * from './fileFields';
+export * from './fileTypes';
 export * from './lineChanges';
 export * from './goal';
 export * from './proposalFields';
