+++ src/shared/schemas/mainView/state.ts
@@ -10,40 +10,18 @@ import {
   UIFileFieldsSchema,
   migrateLegacyContextFileFields,
 } from '../fileFields';
+import {
+  DocumentFileTypeSchema,
+  MultipleDocumentFileTypeSchema,
+} from '../fileTypes';
 import { ToolConfigFieldsSchema } from '../toolConfig';
 
 // ============================================================
-// Session and File Type Schemas
+// Session Schemas
 // ============================================================
 
 export const SessionTypeSchema = z.enum(['toolUse', 'workflow']);
 export type SessionType = z.infer<typeof SessionTypeSchema>;
-
-export const DocumentFileTypeSchema = z.enum(['input', 'context', 'media']);
-export type DocumentFileType = z.infer<typeof DocumentFileTypeSchema>;
-
-export const MultipleDocumentFileTypeSchema = z.enum([
-  'input',
-  'context',
-  'media',
-  'output',
-]);
-export type MultipleDocumentFileType = z.infer<
-  typeof MultipleDocumentFileTypeSchema
->;
-export const MULTIPLE_DOCUMENT_FILE_TYPES =
-  MultipleDocumentFileTypeSchema.options;
-
-export const ExtendedDocumentFileTypeSchema = z.enum([
-  'input',
-  'context',
-  'media',
-  'edited',
-  'output',
-]);
-export type ExtendedDocumentFileType = z.infer<
-  typeof ExtendedDocumentFileTypeSchema
->;
 
 // ============================================================
 // Option Data Schemas
